Data types

+++ Models/ActivitySaveRequestModel.cs
@@ -10,13 +10,13 @@ namespace WaitTime.Models
         public string Name { get; set; }
         public DateTimeOffset StartDateTime { get; set; }
         public DateTimeOffset EndDateTime { get; set; }
-        public int TotalTimeSeconds { get; set; }
-        public int SkiTimeSeconds { get; set; }
-        public int VerticalMeters { get; set; }
-        public int MaxAltitudeMeters { get; set; }
-        public float DistanceMeters { get; set; }
-        public float TopSpeedMps { get; set; }
-        public float AverageSpeedMps { get; set; }
+        public double TotalTimeSeconds { get; set; }
+        public double SkiTimeSeconds { get; set; }
+        public double VerticalMeters { get; set; }
+        public double MaxAltitudeMeters { get; set; }
+        public double DistanceMeters { get; set; }
+        public double TopSpeedMps { get; set; }
+        public double AverageSpeedMps { get; set; }
         public short RunsCount { get; set; }
         public Guid UserId { get; set; }
         public string Source { get; set; }
